requirements.txt
aiogram~=3.0.0b7
python-dotenv~=1.0.0
emoji~=2.2.0
requests~=2.28.2